+++ api/src/models/types.ts
@@ -1,3 +1,5 @@
+import { WebSocket as WSWebSocket } from 'ws';
+
 export interface AquaTemplatesFields {
  
     name: string,
@@ -6,3 +8,11 @@ export interface AquaTemplatesFields {
     required: boolean,
     isArray: boolean
   }
+
+
+  // Interface for client connection with user ID
+export interface ClientConnection {
+  socket: WSWebSocket;
+  userId: string;
+  connectedAt: Date;
+}
