+++ api/src/server.ts
@@ -24,18 +24,10 @@ import fetchChainController from './controllers/fetch-chain';
 import templatesController from './controllers/templates';
 import { setUpSystemTemplates } from './utils/api_utils';
 import systemController from './controllers/system';
-import webSocketController, {broadcastToAllClients} from './controllers/websocket';
-import { WebSocket as WSWebSocket } from 'ws';
+import webSocketController, {broadcastToAllClients} from './controllers/websocketController';
 
-// Store connected WebSocket clients
-export const connectedClients: Map<string, ClientConnection> = new Map();
 
-// Interface for client connection with user ID
-export interface ClientConnection {
-    socket: WSWebSocket;
-    userId: string;
-    connectedAt: Date;
-}
+
 
 
 function buildServer() {
