+++ api/src/store/store.ts
@@ -1,0 +1,5 @@
+import { ClientConnection } from "src/models/types";
+
+// Store connected WebSocket clients
+export const connectedClients: Map<string, ClientConnection> = new Map();
+
